Drop OpenSSL for rustls; shrink artifacts size (#7)
* Drop OpenSSL for rustls

This simplifies onboarding because it drops a built-by-hand dependency
for a regular cargo one. It also required some environment variables,
which are no longer necessary.

It also reduce "bytes still reachable" in valgrind from ~100k to ~10k
when used by the PHP profiler. These bytes appear to mostly be pseudo-
static globals, but there were so many allocations it is difficult to
tell.

I think overall this is a good change, but it does increase the
size of my final PHP executable from 67M to 84M. We need to spend
some time trying to bring that down soon.

* Use line tables only for release builds

Full debug info doesn't seem necessary for getting backtraces.
This shrinks the size back down a bit.

* - Send short commit SHA to libddprof-build
The libddprof-build needs this info to store files in s3.

Co-authored-by: r1viollet <erwan.viollet@datadoghq.com>