Adjust per-platform link libraries (#12)
* Adjust per-platform link libraries

While investigating a sigsegv due to strange linking, I realized that
the link libraries that Rust suggests are per platform. This commit
adjusts the pkg-config and cmake helpers in a per-platform way.

Note that this is brittle. Any change in these flags should be
investigated by a human.

* Remove -lgcc_s for Linux platforms

This is because if it is present, even if -static-libgcc is provided,
then it will be present in the final library runtime dependencies.
This is undesirable at least on Alpine Linux, where it may not even be
present due to the platform's minimalism.