Avoid boxing error for socket_path_to_uri
We should probably do a pass on this at some point, but consumers of
the Rust API should not have to deal with a boxed error generally so
that we avoid allocations and retain information about errors. We
would still have boxed errors in a few cases, most likely:

 1. The errors on FFI boundaries will need boxed for ABI stability.
 2. Having multiple errors from different crates in the same function.
    Although it's possible to make error types to serve this case, it
    does take effort that's probably better spent elsewhere for now.

Anyway, in this case, there is only one error type and the boxing is
unnecesary for Rust callers.