Move ddprof-ffi into profiling crate as a feature
The feature is not enabled by default and is named
`datadog_profiling_ffi`.

I'm not 100% sure this is actually better. cbindgen maps the feature
into a C define, which requires users to set -DDATADOG_PROFILING_FFI
when compiling. I've adjusted the pkg-config and cmake helpers to do
this.

The upside is that it's more contained, and the artifacts are named
libdatadog_profiling.so instead of libdatadog_profiling_ffi.so. It's
also less lines in the repo.

Update licenses

Format cmake files and add Secure32 on Win

Also move the list of libraries that need linked into the source
cmake instead of the example.

Clean up ffi examples a bit