Report profiling data in v2.4 intake format
This was tested with both the Ruby (agent and agentless modes) and
the PHP profilers.

This also introduces a breaking API change: the
`ddog_ProfileExporter_build` / `ProfileExporter::build` functions
now take two additional arguments -- the profiling library name
and version.

Other than that change, using the v2.4 intake format is transparent
to the libdatadog users.

Thanks to @morrisonlevi for pairing with me on this.

Note that this does not (yet) include support for including
attributes in the reporting data. I'll leave that for a separate PR.