IPC framework based on tarpc  (#20)
* IPC transport implementation

Includes the ability to transfer file descriptors between processes, and
provides both an async (tokio based) interface as well as synchronous
interface

* Fix compilation on windows, separate tests that require isolation into
"ignored" tests group

# What does this PR do?
Implement a msgpack based RPC / IPC. Where a server is spawned in an async context, and clients can be either async or sync. In addition the RPC can be used in a 1 way style. Where messages are simply sent onto a pipe, without requiring reading of any responses. 

This way we can optimize for cases where data transmission is simply in a single direction. 

This RPC framework uses `tarpc` crate and adds to it possibility to submit RPC requests without using tokio, and also transport FileDescriptors that can be used to safely delegate processing to the sidecar process.

See example use here:
https://github.com/DataDog/libdatadog/pull/20/files#diff-b5cbb946eecfad576b88206a0a05bf1a78331d41fcd5842db73e9b902831a202R26

Some quick benchmarks
<pre>
<font color="#26A269">write only interface</font>    time:   [1.8545 µs <b>1.9404 µs</b> 2.0843 µs]                                  
                        change: [-0.9627% +12.268% +31.181%] (p = 0.13 &gt; 0.05)
                        No change in performance detected.
<font color="#A2734C">Found 4 outliers among 100 measurements (4.00%)</font>
  4 (4.00%) high severe

<font color="#26A269">two way interface</font>       time:   [18.007 µs <b>19.087 µs</b> 20.356 µs]                               
                        change: [-15.207% -4.5682% +5.7143%] (p = 0.47 &gt; 0.05)
                        No change in performance detected.
<font color="#A2734C">Found 11 outliers among 100 measurements (11.00%)</font>
  6 (6.00%) low mild
  4 (4.00%) high mild
  1 (1.00%) high severe
</pre>

# Motivation

In some cases the context of data we process needs to live longer than a os process that produces it. For these cases we can spawn a subprocess that will receive the data from producers - and later transform and send to the backend/agent.

For efficient operation we need an optimized RPC, and a method to coordinate sidecar lifecycle.