Minor: Document expected result of `ddog_prof_Profile_add`
**What does this PR do?**:

Documents the expected result of the `ddog_prof_Profile_add` API.

**Motivation**:

While validating #80 on Ruby, I called `add` in a way that should've
triggered an error, but saw none.

On closer inspection, I had no error handling for `add` because it
wasn't clear to me that its result could indicate an error.

Thus, I decided to document that fact.

**Additional Notes**:

At some point we could change `add` to return an actual result,
allowing us to pass the error message back to the caller, but for now
I decided to just document it.

**How to test the change?**:

Documentation change only.