Fix CI in main branch by running cargo fmt
**What does this PR do?**:

This PR uses `rustup run nightly cargo fmt --all` to reformat the
code in the main branch and hopefully fix up the failed CI jobs which
are complaining about formatting.

For reference:
* https://gitlab.ddbuild.io/DataDog/apm-reliability/libddprof-build/-/jobs/288056193
* https://gitlab.ddbuild.io/DataDog/apm-reliability/libddprof-build/-/jobs/288056194

**Motivation**:

Bring `main` back to green.

**Additional Notes**:

Having blocking steps in CI for running the formatter seems a bit
too heavy-handed...

**How to test the change?**:

Validate that CI is back to green.