Get rid of atty dependency (#190)
**What does this PR do?**:

This PR gets rid of the `atty` dependency by upgrading our version of
the `criterion` crate (used for microbenchmarks) from version 0.4 to
0.5.1.

**Motivation**:

We want to get rid of `atty` as it was flagged as dependabot as
potentially problematic. In practice we probably wouldn't have been
affected, but I see no reason not to upgrade `criterion`.

**Additional Notes**:

N/A

**How to test the change?**:

Validate that `cargo bench` still compiles and runs fine (which I
did locally).