Minor: Remove Ruby workaround for upstream issue (#193)
**What does this PR do?**:

In our Ruby packaging, we use the `standard` gem, which provides a
code auto-formatter. This gem pulls in `standard-performance` as one
of its dependencies.

Recently, version 1.1.1 of `standard-performance` was released, and
this update was picked up and was breaking in CI.

I reported this issue upstream on
https://github.com/standardrb/standard-performance/issues/11 ;
as well as forcing an earlier version to be used.

Now, 1.1.2 has been released and thus we can remove our workaround
that was avoiding the problematic version.

**Motivation**:

Keep using the latest version of `standard` and its dependencies.

**Additional Notes**:

N/A

**How to test the change?**:

Validate that CI is still green!