refactor internal types
This continues pulling out things from src/profiling/mod.rs and puts
them into `internal::*`. Additionally, it introduces traits for a few
concepts that were already there, just unnamed (and not tidy):

- Id, e.g. FunctionId, MappingId, LocationId, StackTraceId, etc.
- Item, e.g. Function, Mapping, Location, StackTrace, etc.
- PprofItem, e.g. Function, Mapping, Location, etc but not StackTrace.

The DedupExt trait was simplified, which is partly possible because I
understand Rust better now, but also because of the new traits. It
enforces the max-size and will panic if the containers get too big.
Previously, only mappings checked for being full, which is a bit
ironic because it's the least likely in practice to overflow.

Hopefully this explains why it's such a large diff: all these things
are connected, although I probably snuck in a change or two that didn't
strictly have to be in there (it's hard when it's this big to notice).