Bump libdatadog version to 5.0.0 in preparation for release
**What does this PR do?**

This PR bumps the libdatadog version to 5.0.0. This PR looks a bit
different from previous PRs to bump libdatadog version
[(here's the 4.0.0 PR as an example)](https://github.com/DataDog/libdatadog/pull/236)
because in [this PR](https://github.com/DataDog/libdatadog/pull/247)
we've centralized the version on the single `Cargo.toml` file.

Furthermore, we're going from 4.0.0 to 5.0.0 because there were a number
of backwards-incompatible changes to the profiling APIs.

**Motivation:**

Release libdatadog 5.5.0.

**Additional Notes:**

If I haven't missed anything, the backwards incompatible API changes
were the following:

* The value of the `end_timestamp_ns` label is now provided as a
  regular argument to `ddog_prof_Profile_add`
* The libdatadog 5 serializer outputs compressed pprof files
* The exporter has a new API that takes two lists, a list of
  files to compress and a list of files to assume are compressed when
  exporting
* The libdatadog 5 serializer now resets profiles as part of serializing them
* The `ddog_prof_Profile_new` now returns a result structure

**How to test the change?**

I've tested the libdatadog 5 releases using the Ruby profiler, see
https://github.com/DataDog/dd-trace-rb/pull/3169 for my draft PR.