Allow verify trace-protobuf to continue on error, try 2
**What does this PR do?**

In #275, I've attempted to add the
`mainmatter/continue-on-error-comment` github action which replaces a CI
error with a comment, so that it's less annoying when
"Verify trace-protobuf" fails and we don't get a ugly red cross on every
PR.

My first attempt didn't work! I missed that I needed to change the
trigger for the verify action to be on pull requests, not on pushes.

This PR fixes that, and we should see the action comment show up
as an example.

**Motivation:**

Fix #275.

**Additional Notes:**

N/A

**How to test the change?**

See the comment that will show up!