PoC: Reduce memory footprint and allocations for profiling timeline data
**What does this PR do?**

This PR is a proof-of-concept for reducing:

* the number of allocations needed to store profiling timeline data
* the memory space used to store profiling timeline data

It does so by taking advantage of the following observation: the
timeline data is write-only until it's time to serialize.

Thus, instead of keeping a vec of `TrimmedTimestampedObservation`
(which internally keeps another vec), I've decided to just stream
all the bytes to an lz4 compressor, and to make sure to size its
output buffer so as to avoid growing in very small steps.

At serialization time, we do the reverse, sample by sample, and then
immediately write such samples as pprof, reducing the memory footprint
needed even during serialization.

The objective of this PR is to start a discussion on this approach;
the code itself is an awful hack and it's probably better that we
reimplement this idea from scratch if we do like it.

Note that while I cheated on some of the libdatadog tests (because
hey used internal details that changed), this branch passes the full
Ruby profiler test suite and produces -- as far as I could tell --
valid profiles.

**Motivation:**

One of our Ruby on Rails test apps was experiencing heavy memory
bloat when timeline profiling was enabled.

I'm still doing an (internal) writeup with more details, but TL;DR
it's a combination of many threads (100), and glibc creating
a memory arena for each thread, and then libdatadog's use of
malloc combined with Ruby's use of malloc would cause every one of
these arenas to grow a lot.

Here's some quick numbers (more details on the internal writeup)
for a reproducer app I wrote that from my experiments seems to
mimic the behavior of the full Ruby app close enough:

|             Configuration | RSS usage | Note                                                                                                       |
|--------------------------:|----------:|------------------------------------------------------------------------------------------------------------|
|   Baseline (no profiling) |  215 552k | Baseline here includes the profiler being instanced, but never sampling. So it's higher than in real apps. |
| Libdatadog 5, no timeline |  214 396k |                                                                                                            |
|    Libdatadog 5, timeline |  298 640k | Does not stabilize after simulating 15 mins of profiling (e.g. RSS still growing)                          |
|       Prototype, timeline |  223 468k | RSS oscilates between 219 and 223 MiB                                                                      |

**Additional Notes:**

N/A

**How to test the change?**

I'll provide instructions to reproduce my experiments soon.