feat: less mem per entry in the string table
This uses a `Box<str>` instead of `String`, saving 1 `usize` per
entry. When it's serialized to pprof proto, it gets encoded directly
without allocating a temporary `String`. This currently is waiting
on a PR in prost to expose the needed function to directly encode
`AsRef<str>` instead of `&String`. This part enables future
optimizations like using a memory arena for string table data.

PROF-9125