[PROF-8944] Bump libdatadog version to 6.0.0 in preparation for release
**What does this PR do?**

This PR bumps the libdatadog version to 6.0.0 in preparation for
release.

We're going from 5.0.0 to 6.0.0 because of a small, yet still
backwards-incompatible change in the profiling APIs (see below).

**Motivation:**

Release libdatadog 6.0.0.

**Additional Notes:**

As far as I could spot, the only backwards incompatible API change is:

* `ProfileExporter::build`/`ddog_prof_Exporter_Request_build` now take
  an extra (optional/can be `NULL`) argument for attaching system
  information [#289](https://github.com/DataDog/libdatadog/pull/289)

**How to test the change?**

I've tested the libdatadog 6 release locally using the Ruby profiler,
see datadog/dd-trace-rb#3455 for my draft PR.