[PROF-9257] Fix incorrect platform detection for x86_64-linux-gnu/aarch64-linux-gnu
**What does this PR do?**

This PR fixes platform detection not working correctly when the
user's platform is `x86_64-linux-gnu`/`aarch64-linux-gnu`.

This manifested to a user as the following error:

> Profiling was requested but is not supported, profiling disabled: Your
> ddtrace installation is missing support for the Continuous Profiler
> because the `libdatadog` gem installed on your system is missing
> binaries for your platform variant.
> (Your platform: `x86_64-linux-gnu`)
> (Available binaries: `x86_64-linux`, `x86_64-linux-musl`)

In the past, rubygems did this normalization for us (we even had
a comment mentioning it in the code!) but this was removed in
https://github.com/rubygems/rubygems/pull/5852 for Rubygems 3.4+

I'm also bumping the `GEM_MAJOR_VERSION` in the `version.rb` file
so I can release this fix as libdatadog 6.0.0.2.0 for Ruby as
soon as this PR is merged.

**Motivation:**

Make sure our platform detection code is correct.

**Additional Notes:**

N/A

**How to test the change?**

This change includes test coverage.