[NO-TICKET] Add validation for Ruby gem file permissions
**What does this PR do?**

This PR adds an additional file permission validation step while
building the libdatadog Ruby packages.

This is effectively the same as
https://github.com/datadog/dd-trace-rb/pull/3531 , and I'm adding it
to avoid the same potential issue.

**Motivation:**

Avoid issue we ran into with `ddtrace` packaging.

**Additional Notes:**

N/A

**How to test the change?**

Running `bundle exec rake package` will trigger the packaging of
the Ruby packages, and you'll see the test running. To see it failing,
`chmod` the `.rb` files to something other than 644.