[NO-TICKET] Package libdatadog v7.0.0 for Ruby
**What does this PR do?**

This PR includes the changes documented in the "Releasing a new version
to rubygems.org" part of the README:
https://github.com/datadog/libdatadog/tree/main/ruby#releasing-a-new-version-to-rubygemsorg

(It's also very similar to the same as the v6.0.0 release PR).

**Motivation:**

Enable Ruby to use libdatadog v7.0.0.

**Additional Notes:**

N/A

**How to test the change?**

I've tested this release locally against the changes in
https://github.com/DataDog/dd-trace-rb/pull/3536 .

As a reminder, new libdatadog releases don't get automatically picked up
by dd-trace-rb, so the PR that bumps the Ruby profiler will also test
this release against all supported Ruby versions.