[PROF-8917] Add Ruby helper to get `ld_library_path` for libdatadog
**What does this PR do?**

This PR adds a `Libdatadog.ld_library_path` helper to the libdatadog
Ruby support code.

This helper returns the full path to the `lib/` folder where the
libdatadog `.so` files are contained.

**Motivation:**

I'm adding this to support the crashtracker for Ruby: we need this
information when starting the crashtracker, and rather than having
the knowledge of folder layouts inside dd-trace-rb, I think it
makes sense for the libdatadog Ruby code to keep this knowledge.

**Additional Notes:**

I'm planning to include this change in the libdatadog 9 release
for Ruby.

**How to test the change?**

Test coverage is included (and CI runs these tests).