refactor(profiling): backup sample types/period (#403)
This breaks a dependency on the string table implementation details to
be able to reset a profile and preserve the sample types and period. It
does cause more allocations, though.

Adds api::ValueType::new and uses it instead of constructing it
directly, because in most cases with `new` it becomes a single line
instead of four.