Fix telemetry shutdown and log max filter
Fixing max() in max_level_hint().

And properly dropping the telemetry worker:
When the runtime is shutdown with pending apps, it will send Stop, but never actually shutdown the telemetry instances. This regressed with d1fb3bc934e4b51b2ba1cb45818717351ad43b0.
Now we simply drop the telemetry handle, which implicitly stops the worker as well causes the worker to be joined properly.

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>