Fix coalescing send data
There's an interesting tidbit of information in the documentation of dedup_by:

> The elements are passed in opposite order from their order in the slice, so if same_bucket(a, b) returns true, a is removed.

So, we were adding the data to a.tracer_payloads ... and then a gets dropped. Awesome, we just successfully dropped everything but one value.

On top of that, have a size restriction for the possible payload size, to ensure we'll never exceed it.

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>