Add option to skip checks when building profiling
Adds -T option to build-profiling-ffi.sh script to skip
post-build checks, which is frequently done during development.

Also adds a usage blurb which can be printed with -h
option, and prints the usage blurb if the required
positional parameter of destination directory is not provided
instead of failing with a non-descriptive error.