[NO-TICKET] Validate that all public symbols are prefixed in Ruby releases
**What does this PR do?**

This PR adds a check during packaging of Ruby releases that all symbols
exported in the libdatadog shared library are correctly prefixed.

**Motivation:**

This allows us to catch any new unprefixed symbols, and thus we're
able to rely on libdatadog not exporting symbols that would clash with
other libraries that may be loaded in the same Ruby VM.

**Additional Notes:**

N/A

**How to test the change?**

You can observe the new check running with
`docker-compose run push_to_rubygems`.

This is safe to call as it'll stop asking for credentials before
actually publishing anything to rubygems.