Retry when poll_read_ready immediately succeeds in AsyncChannel
UnixStream::poll_read_ready may return Poll::Ready(Ok(())), which incidentally, when returned by an AsyncRead::poll_read function without further modification of buf, means that the channel was closed.
The intended meaning is "immediately available", and thus we need to re-read.

This only can happen when epoll() is resumed between a recvmsg() call and the poll_ready_ready function being invoked, which may only happen in a multithreaded environment, which we recently switched to.

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>