[PROF-10656] Add crash_address to crashtracker reports
**What does this PR do?**

While investigating a customer crash, I found that I dearly missed
having the address that caused a segfault.

E.g. given this code:

```c
  char *potato = 0x1234;
  fprintf(stderr, "Value of potato is %c\n", *potato);
```

...I want to know that the crash was triggered when 0x1234 was
dereferenced, as this information may be key in debugging the issue.

We already get this information in our signal handler, so this PR
just adds the needed piping to report it to datadog.

**Motivation:**

Improve debugging of crashes reported by crashtracker.

**Additional Notes:**

I added the `crash_address` as a tag, similar to `siginfo` and
`signame`. We may want to send it as some other thing; no strong
opinions there.

**How to test the change?**

I was able to test this with the Ruby profiler. I've been trying
to update the tests but there's still some failing and I'm kinda
struggling with that. Hopefully someone with a bit more rust-fu than me
can help out there.